refactor: Sliced views.py into separate files
